Fix interpolation method for arbitrary-point data
I have replaced the old interpolation method (linear interpolation) for generating synthetic data at arbitrary points (not necessarily grid points) with a new one. Note that the new one assumes the time instant we interpolate data at is a multiple of the step size (i.e., it is one of the elements in the vector `T`). If this condition is violated at runtime, an error will be thrown.